+++ custom_components/printer_control_center/dashboards.py
@@ -53,6 +53,10 @@ _DEFAULT_DASHBOARDS: tuple[dict[str, Any], ...] = (
                         {
                             "type": "grid",
                             "cards": [
+
+                                {
+                                    "type": "custom:printer-control-center-camera-card",
+                                },
                                 {
                                     "type": "custom:printer-control-center-queue-card",
                                     "card_size": "xl",